hosts: fix gpg-agent when used on a system with Gnome
When using Gnome as the desktop environment, the Gnome Keychain gets
installed by default. The keychain registers itself as the ssh-agent. I
want to use gnupg-agent for this instead. Some Gnome applications rely
on the keychain, so disabling is not desired. Therefore the ssh-agent
functionality is disabled using a meson build flag though an overlay.